Fail if coords are outside of a safe range
The computation of sentinel edges in InitEdgeDict potentially overflows
if coordinates are more than half of float max/min, and it distinguishes
between real coord bounds and those sentinel edges by a 0.01 margin,
which is not distinguishable at that magnitude due to loss-of-precision.
That can cause assert failures in debug mode or uninitialized reads or
crashes in optimized mode.

Instead, fail if coordinates are not in `[-2^32, 2^32]`. This defines
TESS_MAX_VALUE and TESS_MIN_VALUE in the public headers as those bounds,
in case clients want to clamp or otherwise sanitize invalid inputs.