Another fix for #22
tesedgeSign() does not evaluate to the same sign as tesedgeEval() when the x-coordinates are almost zero. In the test data we had values like -1.70990830e-08f, -8.38190317e-09f, -3.35276112e-10f.

The fix is to always use tesedgeEval() so that we have consistent signs.

This is slight perf degrade, but the code written in '94 tried really hard to avoid one floating point division, and I think we can handle with that today.