Fail if coords are outside of a safe range (#71)
The sentinel edges in `InitEdgeDict` are separated from real edges by the width/height of the bounding box, so this overflows if the dimensions of that box are half of float max.

Instead, fail if coordinates are not in `[-2^32, 2^32]` to avoid overflow and loss-of-precision errors. This `#define`s TESS_MAX_VALID_INPUT_VALUE and TESS_MIN_VALID_INPUT_VALUE in the public headers as those bounds, in case clients want to clamp or otherwise sanitize invalid inputs.